# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrapMaster API — F.S.H.C.L. Sektioun Wolz
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /spieler:
    get:
      operationId: listSpieler
      tags: [spieler]
      summary: List all players
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: object
                properties:
                  spieler:
                    type: array
                    items:
                      $ref: "#/components/schemas/SpielerProfile"
                required: [spieler]
...
  /sync/status:
    get:
      operationId: getSyncStatus
      tags: [sync]
      summary: Connectivity test for ESP32
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  spielerCount:
                    type: integer
                required: [status, timestamp, spielerCount]
  /sync/spieler:
    get:
      operationId: getSyncSpieler
      tags: [sync]
      summary: Get player list for ESP32
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: Player list
          content:
            application/json:
              schema:
                type: object
                properties:
                  spieler:
                    type: array
                    items:
                      $ref: "#/components/schemas/SyncSpieler"
                required: [spieler]
    post:
      operationId: postSyncSpieler
      tags: [sync]
      summary: Sync player list from ESP32
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                spieler:
                  type: array
                  items:
                    $ref: "#/components/schemas/SyncSpieler"
              required: [spieler]
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced:
                    type: integer
                required: [synced]
  /sync/spiele:
    post:
      operationId: postSyncSpiele
      tags: [sync]
      summary: Upload game results from ESP32
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                spiele:
                  type: array
                  items:
                    $ref: "#/components/schemas/SyncSpiel"
              required: [spiele]
      responses:
        "200":
          description: Sync results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        externalId:
                          type: string
                        status:
                          type: string
                          enum: [created, skipped]
                      required: [externalId, status]
                required: [results]

  # ── PRODUCTS & SALES ─────────────────────────────────────────────────────
...
  /sync/products:
    get:
      operationId: getSyncProductCatalog
      tags: [sync]
      security: [{ apiKeyAuth: [] }]
      responses:
        "200": { description: Active catalogue, content: { application/json: { schema: { $ref: "#/components/schemas/ProductList" } } } }
  /sync/sales:
    get:
      operationId: getSyncDaySales
      tags: [sync]
      security: [{ apiKeyAuth: [] }]
      parameters: [{ name: datum, in: query, required: true, schema: { type: string, format: date } }]
      responses:
        "200": { description: Day aggregates, content: { application/json: { schema: { $ref: "#/components/schemas/DaySalesReport" } } } }
    post:
      operationId: postSyncSaleEvents
      tags: [sync]
      security: [{ apiKeyAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SaleEventBatchInput" } } }
      responses:
        "200": { description: Idempotent upload result, content: { application/json: { schema: { $ref: "#/components/schemas/SaleSyncResult" } } } }
        "400": { description: "Unknown player, product, or price revision", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }

# ── COMPONENTS ────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ProductCategory:
      type: string
      enum: [GAME_CREDIT, AMMO_CAL12, AMMO_CAL20, FOOD, DRINK]
    Product:
      type: object
      required: [id, name, category, isSystem, active]
      properties:
        id: { type: integer }
        code: { type: [string, "null"] }
        name: { type: string }
        category: { $ref: "#/components/schemas/ProductCategory" }
        isSystem: { type: boolean }
        active: { type: boolean }
        currentPrice: { anyOf: [{ $ref: "#/components/schemas/PriceRevision" }, { type: "null" }] }
    ProductList:
      type: object
      required: [products]
      properties:
        products: { type: array, items: { $ref: "#/components/schemas/Product" } }
...
    PriceRevision:
      type: object
      required: [id, productId, unitPriceCents, effectiveFrom]
      properties:
        id: { type: integer }
        productId: { type: integer }
        unitPriceCents: { type: integer }
        effectiveFrom: { type: string, format: date-time }
    SaleEventInput:
      type: object
      required: [externalId, spielerId, datum, productId, priceRevisionId, quantity]
      properties:
        externalId: { type: string, minLength: 1 }
        spielerId: { type: integer, minimum: 1 }
        datum: { type: string, format: date }
        productId: { type: integer, minimum: 1 }
        priceRevisionId: { type: integer, minimum: 1 }
        quantity: { type: integer, not: { const: 0 } }
    SaleEventBatchInput:
      type: object
      required: [events]
      properties:
        events: { type: array, items: { $ref: "#/components/schemas/SaleEventInput" } }
    SaleSyncResult:
      type: object
      required: [synced, skipped]
      properties:
        synced: { type: integer }
        skipped: { type: integer }
    DaySalesItem:
      type: object
      required: [productId, quantity, totalCents]
      properties:
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        totalCents: { type: integer }
    DaySalesReport:
      type: object
      required: [datum, sales, totalCents]
      properties:
        datum: { type: string, format: date }
        sales: { type: array, items: { $ref: "#/components/schemas/DaySalesItem" } }
        totalCents: { type: integer }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Modus:
      type: string
      enum:
        - NORMAL
        - HARAKIRI
        - HARAKIRI_DELAYED
        - HARAKIRI_FULL
        - CUSTOM_1
        - CUSTOM_2
        - CUSTOM_3
    Maschine:
      type: string
      enum: [A, B, C, D, E, F, G, H]
...
    SpielerProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        mitgliedNr:
          type: string
          nullable: true
        portalAktiv:
          type: boolean
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, portalAktiv, isAdmin, createdAt]
...
    SyncSpieler:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mitgliedNr:
          type: string
          nullable: true
      required: [id, name]
    SyncSpiel:
      type: object
      properties:
        externalId:
          type: string
          format: uuid
        datum:
          type: string
          format: date-time
        modus:
          $ref: "#/components/schemas/Modus"
        lauf:
          type: integer
          minimum: 1
          maximum: 2
        abgeschlossen:
          type: boolean
        teilnahmen:
          type: array
          items:
            $ref: "#/components/schemas/SyncTeilnahme"
        ergebnisse:
          type: array
          items:
            $ref: "#/components/schemas/SyncErgebnis"
      required: [externalId, datum, modus, lauf, abgeschlossen, teilnahmen, ergebnisse]
    SyncTeilnahme:
      type: object
      properties:
        spielerId:
          type: integer
        startPosten:
          type: integer
          minimum: 1
          maximum: 6
        punkte:
          type: integer
          minimum: 0
          maximum: 36
        lauf:
          type: integer
          minimum: 1
          maximum: 2
      required: [spielerId, startPosten, punkte, lauf]
    SyncErgebnis:
      type: object
      properties:
        spielerId:
          type: integer
        lauf:
          type: integer
          minimum: 1
          maximum: 2
        taube:
          type: integer
          minimum: 1
          maximum: 9
        maschine:
          $ref: "#/components/schemas/Maschine"
        posten:
          type: integer
          minimum: 1
          maximum: 6
        schuss1:
          type: boolean
        schuss2:
          type: boolean
        punkte:
          type: integer
          minimum: 0
          maximum: 4
        wiederholt:
          type: boolean
      required: [spielerId, lauf, taube, maschine, posten, schuss1, schuss2, punkte, wiederholt]