# lib/api-zod/src/generated/types/saleEventBatchInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrapMaster API — F.S.H.C.L. Sektioun Wolz
 * OpenAPI spec version: 0.1.0
 */
import type { SaleEventInput } from './saleEventInput';

export interface SaleEventBatchInput {
  events: SaleEventInput[];
}