# lib/api-zod/src/generated/types/saleEventInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrapMaster API — F.S.H.C.L. Sektioun Wolz
 * OpenAPI spec version: 0.1.0
 */

export interface SaleEventInput {
  /** @minLength 1 */
  externalId: string;
  /** @minimum 1 */
  spielerId: number;
  datum: Date;
  /** @minimum 1 */
  productId: number;
  /** @minimum 1 */
  priceRevisionId: number;
  quantity: number;
}