# lib/api-zod/src/generated/types/saleSyncResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrapMaster API — F.S.H.C.L. Sektioun Wolz
 * OpenAPI spec version: 0.1.0
 */

export interface SaleSyncResult {
  synced: number;
  skipped: number;
}